refactor: move the join-reordering machinery under the rule that owns it
The join graph, its statistics and cost-estimator units sat in
Optimizer/Utils, the directory for helpers shared between optimizer rules.
None of them is shared: every one has zero production consumers outside
optimize-join-order, where the genuinely shared helpers in that directory
have two to four each. They are the internals of a single rule.

Move all five units into a subdirectory owned by that rule. No code
changes -- only the eight files' location and the include paths that name
them.
