+++ arangod/Aql/Optimizer/Rule/CMakeLists.txt
@@ -68,3 +68,5 @@ target_sources(arango_aql PRIVATE
   UseIndexes.cpp
   UseVectorIndex.cpp
 )
+
+add_subdirectory(OptimizeJoinOrder)
