+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(arango_aql PRIVATE
+  IndexJoinStatistics.cpp
+  JoinGraph.cpp
+  SystemRCostEstimator.cpp
+)
