+++ arangod/Aql/Optimizer/Rule/OptimizeJoinOrder/JoinCostEstimator.h
@@ -22,7 +22,7 @@
 
 #pragma once
 
-#include "Aql/Optimizer/Utils/JoinGraph.h"
+#include "Aql/Optimizer/Rule/OptimizeJoinOrder/JoinGraph.h"
 
 #include <algorithm>
 #include <cmath>
