+++ arangod/Aql/Optimizer/Utils/CMakeLists.txt
@@ -1,9 +1,6 @@
 target_sources(arango_aql PRIVATE
   AccessesCollectionVariable.cpp
   EnumeratePathsFilterMatcher.cpp
-  IndexJoinStatistics.cpp
-  JoinGraph.cpp
   LateMaterializedCommon.cpp
   OptimizeTraversalPathVariable.cpp
-  SystemRCostEstimator.cpp
 )
